% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pmtiles2.R
\name{tmap_source_bbox}
\alias{tmap_source_bbox}
\title{Get the bounding box of a tmap source}
\usage{
tmap_source_bbox(meta)
}
\arguments{
\item{meta}{List. Output of \code{\link{tmap_source_meta}}.}
}
\value{
An \code{sf} bounding box (\code{bbox}) with CRS OGC:CRS84.
}
\description{
Convenience wrapper that exposes the bounding box computed inside
\code{\link{tmap_source_meta}}.
}
