% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_pmtiles2.R
\name{tmap_source_layers}
\alias{tmap_source_layers}
\title{List layers in a tmap source}
\usage{
tmap_source_layers(meta)
}
\arguments{
\item{meta}{List. Output of \code{\link{tmap_source_meta}}.}
}
\value{
Character vector of layer names.
}
\description{
Convenience wrapper around \code{\link{tmap_source_meta}} that returns only
the layer names.
}
