% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmap_source.R
\name{tmap_source_meta}
\alias{tmap_source_meta}
\title{Get tmap source meta data}
\usage{
tmap_source_meta(x)
}
\arguments{
\item{x}{url to the source, or file path in case it is stored locally}
}
\value{
list of meta data
}
\description{
Get tmap source meta data
}
